requirements.txt
Django==2.0.2
djangorestframework==3.7.7
pytz==2017.3